+++ lib/farmbot/bootstrap/settings_sync.ex
@@ -32,7 +32,6 @@ defmodule Farmbot.Bootstrap.SettingsSync do
     Map.new(new_map, fn({key, new_value}) ->
       # Logger.debug 1, "Applying #{key} #{inspect old_map[key]} over #{inspect new_value}"
       if old_map[key] != new_value do
-        # Logger.debug 1, "Updating: #{key} => #{new_value}"
         case new_value do
           val when is_boolean(val) ->
             update_config_value(:bool, "settings", key, new_value)
@@ -41,6 +40,7 @@ defmodule Farmbot.Bootstrap.SettingsSync do
           val when is_number(val) ->
             update_config_value(:float, "settings", key, new_value / 1)
         end
+        Logger.success 2, "Updating: #{key} => #{new_value}"
       end
       {key, new_value}
     end)
@@ -59,6 +59,7 @@ defmodule Farmbot.Bootstrap.SettingsSync do
     beta_opt_in = get_config_value(:bool, "settings", "beta_opt_in")
     disable_factory_reset = get_config_value(:bool, "settings", "disable_factory_reset")
     firmware_output_log = get_config_value(:bool, "settings", "firmware_output_log")
+    firmware_input_log = get_config_value(:bool, "settings", "firmware_input_log")
     sequence_body_log = get_config_value(:bool, "settings", "sequence_body_log")
     sequence_complete_log = get_config_value(:bool, "settings", "sequence_complete_log")
     sequence_init_log = get_config_value(:bool, "settings", "sequence_init_log")
@@ -72,6 +73,7 @@ defmodule Farmbot.Bootstrap.SettingsSync do
       beta_opt_in: beta_opt_in,
       disable_factory_reset: disable_factory_reset,
       firmware_output_log: firmware_output_log,
+      firmware_input_log: firmware_input_log,
       sequence_body_log: sequence_body_log,
       sequence_complete_log: sequence_complete_log,
       sequence_init_log: sequence_init_log,
@@ -90,6 +92,7 @@ defmodule Farmbot.Bootstrap.SettingsSync do
     "beta_opt_in",
     "disable_factory_reset",
     "firmware_output_log",
+    "firmware_input_log",
     "sequence_body_log",
     "sequence_complete_log",
     "sequence_init_log",